[EventDispatcher][HttpKernel] Move RegisterListenersPass from HttpKernel to EventDispatcher.
